requirements.txt
fastapi
uvicorn
sentence-transformers
torch
googletrans
langdetect
pymupdf
faiss-cpu
google-generativeai
python-multipart
pydantic
reportlab
